Skew diagram for DNA String.
In this code i use pyplot library for showing the Skew diagram.
Finding skew diagram from DNA string is #G-#C result in a DNA string.We know in DNA string ,their are A,T,G,C.

Which is :
A = adenine,
T = thymine,
C= cytosine,
G = guanine

but in DNA string every time in Replication G(guanine) and C(cytosine) are Mutated  very much.
so, when we want to find Oric in DNA string we need to find the Mutated path of a DNA String.